% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllDependencies.R
\name{addNodeLabels}
\alias{addNodeLabels}
\title{Add Labels to the Nodes Drawn on a Circle Line}
\usage{
addNodeLabels(nodes, cex = 1, distance.factor = 1)
}
\arguments{
\item{nodes}{data frame as returned by \code{\link{toNodes}}}

\item{cex}{character expansion factor as given to \code{text}}

\item{distance.factor}{expansion factor applied to the x and y coordinates of the nodes to get 
the coordinates of the labels}
}
\description{
Add Labels to the Nodes Drawn on a Circle Line
}
