% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anglesToPoints}
\alias{anglesToPoints}
\title{Angle in Unit Circle to Point Coordinate}
\usage{
anglesToPoints(angles.grad)
}
\arguments{
\item{angles.grad}{vector of angles in degree}
}
\value{
matrix with columns \emph{x} and \emph{y} containing the coordinates
  of points corresponding to the given angles in a unit circle
}
\description{
Convert angles in a unit circle to point coordinates (x, y)
}
\examples{
plot(anglesToPoints(equidistantAngles(n = 10)), type = "b")
  
}
