% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedCranVersions.R
\name{archivedCranVersions}
\alias{archivedCranVersions}
\title{archivedCranVersions --------------------------------------------------------
Archived CRAN versions}
\usage{
archivedCranVersions(package, ref_date = NULL)
}
\arguments{
\item{package}{package name}

\item{ref_date}{default: NULL}
}
\description{
archivedCranVersions --------------------------------------------------------
Archived CRAN versions
}
\examples{
packages <- c("ggplot2", "swmmr", "kwb.hantush")
archivedCranVersions(packages)
archivedCranVersions(packages, ref_date= "2012-12-01")
}
