% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{catIf}
\alias{catIf}
\title{Copy of kwb.utils::catIf}
\usage{
catIf(condition, ...)
}
\arguments{
\item{condition}{see \code{kwb.utils::catIf}}

\item{\dots}{see \code{kwb.utils::catIf}}
}
\description{
Copy of kwb.utils::catIf
}
