% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareInstalledVersions.R
\name{compareInstalledVersions}
\alias{compareInstalledVersions}
\title{Compare Package Versions Between Libraries}
\usage{
compareInstalledVersions(lib1, lib2)
}
\arguments{
\item{lib1}{path to first R library}

\item{lib2}{path to second R library}
}
\value{
data frame with columns \code{name} (package name), \code{version.1},
  \code{version.2} (version string of package in \code{lib1} and \code{lib2},
  respectively), \code{differs} (logical indicating whether the version is
  different in the two libraries)
}
\description{
Compare Package Versions Between Libraries
}
