% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyBasePackages.R
\name{copyBasePackages}
\alias{copyBasePackages}
\title{Copy Base R Packages from the System Library to the Target Library}
\usage{
copyBasePackages(
  target_lib,
  set_number = 2L,
  system_lib = utils::tail(.libPaths(), 1L),
  packages = systemPackages(set_number)
)
}
\arguments{
\item{target_lib}{path to the target library}

\item{set_number}{number defining the base packages to be copied, see
\code{\link{systemPackages}}}

\item{system_lib}{path to the system library from which to copy packages}

\item{packages}{vector of names of packages to be copied}
}
\description{
Copy Base R Packages from the System Library to the Target Library
}
