% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranVersions.R
\name{cranVersions}
\alias{cranVersions}
\title{Get versions of CRAN packages}
\usage{
cranVersions(name, dbg = TRUE)
}
\arguments{
\item{name}{package name}

\item{dbg}{logical indicating whether or not to show debug messages. Default: 
\code{TRUE}}
}
\description{
Get versions of CRAN packages
}
