% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadFile}
\alias{downloadFile}
\title{Downloads File from URL and Returns the Path to the Downloaded File}
\usage{
downloadFile(url, targetDir = tempdir(), quiet = TRUE)
}
\arguments{
\item{url}{URL to file to be downloaded}

\item{targetDir}{path to local target directory}

\item{quiet}{passed to \code{\link{download.file}}}
}
\value{
path to downloaded file
}
\description{
Downloads File from URL and Returns the Path to the Downloaded File
}
