% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGitHubPackage.R
\name{downloadGitHubPackage}
\alias{downloadGitHubPackage}
\title{Download an R Package from GitHub}
\usage{
downloadGitHubPackage(repo, destdir = "~/../Downloads")
}
\arguments{
\item{repo}{path to repository, relative to https://github,com, e.g. 
"kwb-r/kwb.utils"}

\item{destdir}{path to download folder, default: "~/../Downloads"}
}
\value{
path to downloaded file in the \code{destdir} folder with attribute 
  "origin" pointing to the original file in \code{tempdir()}.
}
\description{
Download an R Package from GitHub
}
