% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPackagesFromSnapshot.R
\name{downloadPackagesFromSnapshot}
\alias{downloadPackagesFromSnapshot}
\title{Download Package Archive from Microsoft R Archive Network}
\usage{
downloadPackagesFromSnapshot(
  packages,
  snapshot_date,
  destdir = NULL,
  type = c("source", "win.binary")[1L]
)
}
\arguments{
\item{packages}{names of packages (vector of character)}

\item{snapshot_date}{date of snapshot of CRAN package versions, as a 
string in yyyy-mm-dd format}

\item{destdir}{path to download folder}

\item{type}{one of \code{c("source", "win.binary")}}
}
\value{
paths to the downloaded files (vector of character)
}
\description{
Download Package Archive from Microsoft R Archive Network
}
