% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllDependencies.R
\name{drawDependencies}
\alias{drawDependencies}
\title{Draw Lines between Nodes}
\usage{
drawDependencies(nodes, dependencies, nodeColours, ...)
}
\arguments{
\item{nodes}{data frame as returned by \code{\link{toNodes}}}

\item{dependencies}{list of package dependencies as returned by 
\code{\link{packageDependencies}}}

\item{nodeColours}{colours given to the lines starting at the same start node}

\item{\dots}{arguments passed to \code{\link{drawLink}}}
}
\description{
Draw Lines between Nodes
}
