% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{drawLink}
\alias{drawLink}
\title{Draw a Link Between a Pair of Nodes}
\usage{
drawLink(nodes, i, j, ...)
}
\arguments{
\item{nodes}{data frame as returned by \code{\link{toNodes}}}

\item{i}{row index of the start node}

\item{j}{row index of the finish node}

\item{\dots}{arguments passed to \code{arrows}}
}
\description{
Draw a Link Between a Pair of Nodes
}
