% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equidistantAngles}
\alias{equidistantAngles}
\title{Equidistantly Distributed Angles in Degrees Between 0 and 360}
\usage{
equidistantAngles(n, from = 0)
}
\arguments{
\item{n}{number of angles to be returned}

\item{from}{start angle in degrees. Default: 0}
}
\description{
Equidistantly Distributed Angles in Degrees Between 0 and 360
}
\examples{
x <- equidistantAngles(90)

plot(x, sin(gradToRad(x)), xlab = "angle in degree", ylab = "sin(angle)")

}
