% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github_packages.R
\name{get_r_version_majorminor}
\alias{get_r_version_majorminor}
\title{Helper: Get R major minor version string}
\usage{
get_r_version_majorminor()
}
\value{
returns R version major.minor string (e.g. 4.0), used by standard R 
libraries for grouping all R packages into one folder
}
\description{
Helper: Get R major minor version string
}
\examples{
get_r_version_majorminor(
)
}
