% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAuthors}
\alias{getAuthors}
\title{Get Information on Package Authors}
\usage{
getAuthors(package)
}
\arguments{
\item{package}{name of (installed) package}
}
\description{
Get Information on Package Authors
}
