% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentCranVersion.R
\name{getCranPackageDatabase}
\alias{getCranPackageDatabase}
\title{Get Matrix with Information on All CRAN Packages}
\usage{
getCranPackageDatabase()
}
\description{
Get Matrix with Information on All CRAN Packages
}
