% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDependencyData.R
\name{getDependencyData}
\alias{getDependencyData}
\title{Get Package Dependency Data from Package Database}
\usage{
getDependencyData(
  db,
  fields = c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"),
  dbg = FALSE
)
}
\arguments{
\item{db}{package data base as e.g. returned by
\code{\link{getCranPackageDatabase}}}

\item{fields}{types of dependencies to be considered. Default:
\code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}}

\item{dbg}{logical indicating whether or not to show debug messages}
}
\description{
Get Package Dependency Data from Package Database
}
