% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{getPackageFilesToInstall}
\alias{getPackageFilesToInstall}
\title{Get Package Files to Install}
\usage{
getPackageFilesToInstall(package.dir = defaultPackageDir(),
  packageNames = NULL, filepattern = "", full.names = TRUE,
  dbg = FALSE, warn = TRUE)
}
\arguments{
\item{package.dir}{full path to directory containing the package files}

\item{packageNames}{optional vector of character containing the names of the
packages to be installed. If NULL (default), all available packages in
\emph{package.dir} are installed,}

\item{filepattern}{pattern matching names of files to be considered. Default:
"^kwb\\..*\\.zip$"}

\item{full.names}{if TRUE (default) the full paths to the package files are
returned, else only the file names}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}

\item{warn}{if \code{TRUE} (default) warnings are given if no corresponding
package files are found}
}
\value{
character vector containing the full paths to or just the names of
  the available package files
}
\description{
Get paths/names of package files available in a directory
}
