% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageLicences.R
\name{getPackageLicences}
\alias{getPackageLicences}
\title{Which Licences are Specified for the Packages?}
\usage{
getPackageLicences(packages, db = as.data.frame(utils::installed.packages()))
}
\arguments{
\item{packages}{names of (installed) packages}

\item{db}{optional. Package database, similar to what is returned by
\code{\link[utils]{installed.packages}}. Default:
\code{as.data.frame(installed.packages())}}
}
\value{
data frame
}
\description{
Which Licences are Specified for the Packages?
}
