% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getRVersionMajorMinor}
\alias{getRVersionMajorMinor}
\title{Helper: Get R major minor version string}
\usage{
getRVersionMajorMinor()
}
\value{
returns R version major.minor string (e.g. 4.0), used by standard R 
libraries for grouping all R packages into one folder
}
\description{
Helper: Get R major minor version string
}
\examples{
getRVersionMajorMinor()
}
