% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gradToRad}
\alias{gradToRad}
\title{Angle in Degree to Angle in rad}
\usage{
gradToRad(grad)
}
\arguments{
\item{grad}{vector of angles in degrees}
}
\description{
Angle in Degree to Angle in rad
}
\examples{
gradToRad(c(0, 90, 180, 270, 360)) / pi

}
