% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hasGplLicence}
\alias{hasGplLicence}
\title{Do Packages have a GPL Licence?}
\usage{
hasGplLicence(packages)
}
\arguments{
\item{packages}{package name(s) as a vector of character}
}
\value{
vector of logical
}
\description{
Do Packages have a GPL Licence?
}
