% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_cran.R
\name{initLocalCRAN}
\alias{initLocalCRAN}
\title{Create the folder structure for a local CRAN-like repository}
\usage{
initLocalCRAN(local_cran)
}
\arguments{
\item{local_cran}{full path to the folder representing the local CRAN}
}
\description{
Create the folder structure for a local CRAN-like repository
}
