% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installedDependencies.R
\name{installedDependencies}
\alias{installedDependencies}
\title{What Versions of Package Dependencies are Installed?}
\usage{
installedDependencies(package, recursive = TRUE)
}
\arguments{
\item{package}{name of the package of which to check the dependencies}

\item{recursive}{whether to look recursively for dependencies or only for the
direct dependencies of \code{package}. Passed to 
\code{\link{packageDependencies}}, defaults to \code{TRUE}}
}
\description{
What Versions of Package Dependencies are Installed?
}
\examples{
installedDependencies(package = "kwb.package")
installedDependencies(package = "kwb.package", recursive = FALSE)
}
