% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installedKwbPackages}
\alias{installedKwbPackages}
\title{Installed KWB-Packages}
\usage{
installedKwbPackages()
}
\value{
vector of names of installed kwb-packages
}
\description{
Installed KWB-Packages
}
