% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github_packages.R
\name{installGithubPackages}
\alias{installGithubPackages}
\title{Install GitHub Packages}
\usage{
installGithubPackages(
  lib,
  pkgs_full_name,
  dependencies = TRUE,
  upgrade = "always",
  auth_token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{lib}{path to R library where packages should be instaslled}

\item{pkgs_full_name}{vector with full sname to GitHub R packages (e.g. "kwb-r/kwb.utils")}

\item{dependencies}{passed to remotes::install_github(). TRUE is shorthand for 
"Depends", "Imports", "LinkingTo" and "Suggests"  NA is shorthand for "Depends", 
"Imports" and "LinkingTo" and is the default. FALSE is shorthand for no 
dependencies (i.e. just check this package, not its dependencies), (default: TRUE)}

\item{upgrade}{passed to remotes::install_github(), (default: "always")}

\item{auth_token}{GitHub Personal Access token, required with scope "private" if
access to non-public R packages is required (default: Sys.getenv("GITHUB_PAT"))}
}
\value{
installs multiple GitHub R packages into one R library
}
\description{
Install GitHub Packages
}
\examples{
\dontrun{
remotes::install_github("kwb-r/pkgmeta") 
pkgs <- pkgmeta::get_github_packages()
paths_list <- list(
r_version = kwb.packages::getRVersionMajorMinor(),
lib_linux = "/usr/lib/R/site-library",
lib_win = "<win_root_dir>/kwbran/<r_version>"
)

paths <- kwb.utils::resolve(paths_list,
win_root_dir = fs::path(tempdir())) 

pkgs <- pkgmeta::get_github_packages()

installGithubPackages(lib = paths$lib_win, pkgs$full_name)
installGithubPackages(lib = paths$lib_linux, pkgs$full_name)
}

}
