% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{installRemotes}
\alias{installRemotes}
\title{Install the remotes Package to the Given Library}
\usage{
installRemotes(lib)
}
\arguments{
\item{lib}{path to the library to which to install the remotes package}
}
\description{
Install the remotes Package to the Given Library
}
