% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDeps.R
\name{loadDescriptionFromWeb}
\alias{loadDescriptionFromWeb}
\title{Load Package DESCRIPTION from a File in the Internet}
\usage{
loadDescriptionFromWeb(
  name,
  version = NA,
  github_user = "KWB-R",
  path = tempdir(),
  cache = list(descriptions = list(), versions = list())
)
}
\arguments{
\item{name}{package name}

\item{version}{version string. Default: NA}

\item{github_user}{Default: "KWB-R"}

\item{path}{path to local .tar.gz file}

\item{cache}{list with elements "descriptions", "versions", used as a cache}
}
\description{
Load Package DESCRIPTION from a File in the Internet
}
