% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach.R
\name{packageDependencies}
\alias{packageDependencies}
\title{Package Dependencies}
\usage{
packageDependencies(
  packages = NULL,
  db = utils::installed.packages(),
  which = c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")[1:3],
  recursive = TRUE,
  reverse = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{packages}{a character vector of package names.}

\item{db}{character matrix as from \code{\link{available.packages}()}
    (with the default \code{NULL} the results of this call)
    or data frame variants thereof.  Alternatively, a package database
    like the one available from
    \url{https://cran.r-project.org/web/packages/packages.rds}.
  }

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }

\item{recursive}{a logical indicating whether (reverse) dependencies
    of (reverse) dependencies (and so on) should be included, or a
    character vector like \code{which} indicating the type of (reverse)
    dependencies to be added recursively.}

\item{reverse}{logical: if \code{FALSE} (default), regular
    dependencies are calculated, otherwise reverse dependencies.}

\item{verbose}{logical indicating if output should monitor the package
    search cycles.}
}
\description{
This is just a wrapper around \code{\link[tools]{package_dependencies}} with
some defaults defined.
}
