% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_network.R
\name{packageString}
\alias{packageString}
\title{Package String}
\usage{
packageString(package)
}
\arguments{
\item{package}{Package name}
}
\value{
\code{package}, preceded by \code{package:}
}
\description{
Package String
}
