% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDependencies}
\alias{plotDependencies}
\title{Plot Dependencies Between Nodes on a Circle Line}
\usage{
plotDependencies(nodes, dependencies, main = "", r = 1.5,
  nodeColours = grDevices::rainbow(nrow(nodes)), ...)
}
\arguments{
\item{nodes}{data frame as returned by \code{\link{toNodes}}}

\item{dependencies}{list of package dependencies as returned by
\code{\link{packageDependencies}}}

\item{main}{plot title}

\item{r}{radius of the circle}

\item{nodeColours}{colours to be given to the nodes}

\item{\dots}{passed to \code{\link{drawDependencies}}}
}
\description{
Plot Dependencies Between Nodes on a Circle Line
}
