% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNodes}
\alias{plotNodes}
\title{Plot Nodes}
\usage{
plotNodes(nodes, r = 1, col = "red", ...)
}
\arguments{
\item{nodes}{data frame as returned by \code{\link{toNodes}}}

\item{r}{Radius of a circle that fits into the plot range: xlim is set to 
\code{c(-r, r)}}

\item{col}{colour used to plot the nodes (default: "red")}

\item{\dots}{arguments passed to \code{plot}}
}
\description{
Plot Nodes
}
