% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPackageVersions.R
\name{plotPackageVersions}
\alias{plotPackageVersions}
\title{Plot Package Versions}
\usage{
plotPackageVersions(
  versions,
  r_range = c(1, 10),
  rmax = 1.1 * r_range[2L],
  dphi = NULL,
  ticklen = 1
)
}
\arguments{
\item{versions}{versions}

\item{r_range}{r_range (default: c(1, 10))}

\item{rmax}{rmax (default: 1.1 * r_range[2L])}

\item{dphi}{dphi (default: NULL)}

\item{ticklen}{ticklen (default: 1)}
}
\description{
Plot Package Versions
}
