% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_network.R
\name{plotSankeyNetwork}
\alias{plotSankeyNetwork}
\title{Plot Sankey Network}
\usage{
plotSankeyNetwork(functionName, where = 1, ...)
}
\arguments{
\item{functionName}{name of function from which to start the network}

\item{where}{passed to \code{\link[mvbutils]{foodweb}}}

\item{\dots}{additional arguments passed to \code{\link[mvbutils]{foodweb}}}
}
\description{
Plot Sankey Network
}
