% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localCran.R
\name{provideInLocalCRAN}
\alias{provideInLocalCRAN}
\title{Provide a Source Package in the Local Cran}
\usage{
provideInLocalCRAN(
  package,
  rebuild = TRUE,
  local_cran = defaultLocalCRAN(drive_letter = TRUE)
)
}
\arguments{
\item{package}{name of the package to be looked up in either of these
locations: \code{<home>/Documents/R-Development/RPackages}, 
\code{<home>/Desktop/R_Development/RPackages}}

\item{rebuild}{logical. If \code{TRUE} the package is rebuild before all
\code{.tar.gz}-files from the parent folder of the package folder are 
copied to the local CRAN folder structure}

\item{local_cran}{full path to the folder representing the local CRAN}
}
\description{
Provide a Source Package in the Local Cran
}
