% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGithubPackageDescription.R
\name{readGithubPackageDescription}
\alias{readGithubPackageDescription}
\title{readGithubPackageDescription ------------------------------------------------}
\usage{
readGithubPackageDescription(
  repo,
  sha,
  auth_token = remotes_github_pat(),
  destdir = tempdir()
)
}
\arguments{
\item{repo}{GitHub repository, e.g. "kwb-r/kwb.utils"}

\item{sha}{SHA (hash) of the commit}

\item{auth_token}{GitHub token}

\item{destdir}{path to destination directory, i.e. directory to which the
DESCRIPTION file is copied. Default: \code{tempdir()}}
}
\description{
Read DESCRIPTION File for R Package on GitHub
}
