% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localCran.R
\name{setOptionsForPackrat}
\alias{setOptionsForPackrat}
\title{Set Options for Using Packrat}
\usage{
setOptionsForPackrat()
}
\value{
The old options are returned invisibly.
}
\description{
Add the path to the local repository to the option "repos" and set the option
"pkgType" to "source".
}
