% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stopIfNotInstalled}
\alias{stopIfNotInstalled}
\title{Is a Package Installed?}
\usage{
stopIfNotInstalled(package)
}
\arguments{
\item{package}{package name (character vector of length one)}
}
\description{
Is a Package Installed?
}
