% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{systemPackages}
\alias{systemPackages}
\title{Names of Base R Packages}
\usage{
systemPackages(set_number = 1L)
}
\arguments{
\item{set_number}{integer number specifying a set of packages: 1 or 2.}
}
\value{
vector of character representing package names
}
\description{
Names of Base R Packages
}
