% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSankeyNetwork.R
\name{toLinksAndNodes}
\alias{toLinksAndNodes}
\title{Convert Links to List of Links and Nodes}
\usage{
toLinksAndNodes(links)
}
\arguments{
\item{links}{list with elements \code{source} and \code{target}}
}
\value{
list with elements \code{links} (input list \code{links} with new 
  elements \code{value}, \code{source}, \code{target}) and \code{nodes}
  (data frame with column \code{name})
}
\description{
Convert Links to List of Links and Nodes
}
\examples{
kwb.package:::toLinksAndNodes(list(
  source = c("s1", "s1"), target = c("t1", "t2")
))

}
