% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{toNodes}
\alias{toNodes}
\title{Nodes in a Unit Circle}
\usage{
toNodes(nodeNames)
}
\arguments{
\item{nodeNames}{character vector of nodes to be arranged in a unit circle}
}
\value{
data frame with columns \code{x} and \code{y} giving the coordinates
  of the nodes arranged in a circle. The row names represent the node names.
}
\description{
Node names to node coordinates in a unit circle
}
\examples{
nodes <- toNodes(LETTERS)

plot(nodes)
text(nodes, labels = row.names(nodes), adj = c(0, 0))
  
}
