% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{updateKwbPackages}
\alias{updateKwbPackages}
\title{Update or Install KWB-Packages}
\usage{
updateKwbPackages(
  packageNames = sort(installedKwbPackages()),
  skip = character(),
  package.dir = defaultPackageDir()
)
}
\arguments{
\item{packageNames}{vector of packages to be installed. Default: names of all
installed KWB-packages}

\item{skip}{vector of packages not to be installed, even if they are listed
in \emph{packageNames}}

\item{package.dir}{full path to the folder containing the binary package 
files}
}
\description{
Update installed KWB-packages or install KWB-packages for the first time
}
