% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageLicences.R
\name{getPackageLicences}
\alias{getPackageLicences}
\title{Which Licences are Specified for the Packages?}
\usage{
getPackageLicences(packages, stop.on.error = FALSE)
}
\arguments{
\item{packages}{names of (installed) packages}
}
\value{
data frame
}
\description{
Which Licences are Specified for the Packages?
}
