% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubPackageVersions.R
\name{githubVersions}
\alias{githubVersions}
\title{Get Versions of Packages on GitHub}
\usage{
githubVersions(name, github_user = "KWB-R")
}
\arguments{
\item{name}{package name}

\item{github_user}{name of github account, default: "KWB-R"}
}
\value{
data frame with one row per available version
}
\description{
Get Versions of Packages on GitHub
}
\examples{
githubVersions("kwb.utils")
}
