% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach.R
\name{packageDependencies}
\alias{packageDependencies}
\title{Package Dependencies}
\usage{
packageDependencies(packages = NULL, recursive = TRUE,
  reverse = FALSE)
}
\arguments{
\item{packages}{passed to \code{\link[tools]{package_dependencies}}}

\item{recursive}{passed to \code{\link[tools]{package_dependencies}}}

\item{reverse}{passed to \code{\link[tools]{package_dependencies}}}
}
\description{
Package Dependencies
}
