% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach.R
\name{sortedDependencies}
\alias{sortedDependencies}
\title{Sorted Package Dependencies}
\usage{
sortedDependencies(package, dbg = FALSE)
}
\arguments{
\item{package}{name of package of which dependencies are to be found}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
vector of package names. The first element is the package itself,
  followed by the names of depending packages. You should be able to detach
  the packages in this order without any "package ... is required by ..."
  error
}
\description{
Names of depending packages in the order of their dependency
}
